requirements_api.txt
fastapi
numpy
transformers
google-cloud-storage
uvicorn
torch